webrtc_srtp::config

Struct Config

pub struct Config {
    pub keys: SessionKeys,
    pub profile: ProtectionProfile,
    pub local_rtp_options: Option<ContextOption>,
    pub remote_rtp_options: Option<ContextOption>,
    pub local_rtcp_options: Option<ContextOption>,
    pub remote_rtcp_options: Option<ContextOption>,
}

Config is used to configure a session. You can provide either a KeyingMaterialExporter to export keys or directly pass the keys themselves. After a Config is passed to a session it must not be modified.

Fields
keys: SessionKeysprofile: ProtectionProfilelocal_rtp_options: Option<ContextOption>

List of local/remote context options. ReplayProtection is enabled on remote context by default. Default replay protection window size is 64.

remote_rtp_options: Option<ContextOption>local_rtcp_options: Option<ContextOption>remote_rtcp_options: Option<ContextOption>
Implementations

impl Config

pub async fn extract_session_keys_from_dtls( &mut self, exporter: impl KeyingMaterialExporter, is_client: bool, ) -> Result<(), Error>

ExtractSessionKeysFromDTLS allows setting the Config SessionKeys by extracting them from DTLS. This behavior is defined in RFC5764: https://tools.ietf.org/html/rfc5764